requirements.txt
numpy
pandas
matplotlib.pyplot
pandas_datareader
yfinance
streamlit
tensorflow
keras.models
datetime
drive